requirements.txt
exchangelib>=3.3.1,<4.0.0
openpyxl>=3.1.0,<4.0.0
python-dotenv>=0.19.1,<1.0.0
pandas>=1.3.3,<2.0.0
python-dateutil>=2.8.2,<3.0.0
easygui>=0.98.3,<1.0.0
